Issue 1413 - allow marker with `Sound.fadeIn`
`Sound.fadeIn` now accepts a marker as a 3rd parameter. Just as with
`play`, this is where the sound starts at.

This is a minor breaking change; code that was playing from a marker and
then used `fadeIn` expecting to play from the beginning will need to be to
the last parameter explicitly. This does not affect code that always
played a sound from the start.