Removed toc-shim from template files
I wasn't aware in my previous commits that this was all generated by a template. After browsing the issues some more, I found the templates and took out the toc-shim. I don't believe the jQuery scrolling occurs in a template, so the fix that I made to it before should be effectual, but I'm not positive.