Simplifying file list collection.
Now, the build task only have to collect the original source files to be
copied, instead of batching several 'concat' sub-tasks and placing the
intermediate results in a subdirectory in 'dist'.  This cuts the amount of
tasks required for Grunt to build the Phaser distribution files.

Because much of the Phaser modules consists of just single files, taking out
these steps may recude file duplication and the task cycle.

This obsoletes the 'concat' sub-tasks used in the compilation process.