tutorial update
I moved around part of the tutorial. that way if people are trying what
is being said in that section they see an effect. since they added the
update function beforehand.
Otherwise people could wonder ( like myself ) why removing the
immovable of the ground has no effect. of course people could realize
this after reading the entire paragraph but when you code live with the
tutorial this error can occur. and we should make it as easy as possible
for people.