Fixes a bug #3651 Describe the changes below: The tween animation was getting stuck when calling restart on it after being added.
It now checks to see if this.state === TWEEN_CONST.PENDING_ADD in the tween restart function

If true, sets this.seek(0) and this.parent.makeActive(this)

This fixes the issue that makes the tween stuck when restarting the new tween directly after adding a the new tween.

Updates to package.json and package-lock.json were undone from the last pull request.