Update RenderTexture fill
Add parameters to enable filling a rectangle instead of a whole texture. Besides that alpha was ignored for canvas.

I don't know much about gl, so please check if that is ok or if there's any better way.
Please check default values for width and height in @param line, so they don't mess with your documentation generation.