Fix loading audio and video with blob urls
Passing a blob url as the second parameter to `this.load.audio` or
`this.load.video` would cause Phaser to load a different url.
Additionally, `this.load.path` would incorrectly be prepended to the
url.

`AudioFile.getAudioURL` and `VideoFile.getVideoURL`, are expected to
return an object with two properties; url and type. However when passed
a blob url they return a string. The return value's
`url` property is inspected to get the url. For strings' the property is
`undefined`. If the url property is undefined then a url derived from
the asset key is used instead.