official runtime bug
see this:
https://github.com/EsotericSoftware/spine-runtimes/pull/1672
I found this bug when I tried to implement the dyanmic skin switching.

because this is not a bug phaser should maintain, so this PR is I don't kow, reminder? 
so you can merge in advance or wait for official new version and close this PR, up to you.   ;)