Type clarification for emit zone data
I have updated the definition of EmitZoneObject to exclusively represent EdgeZone or RandomZone objects, excluding their configurations. Previously, the emitZones property was either an array of RandomZones or an array of EdgeZones, not an Array that could hold either. With this change, emitZones is now represented as an array of EmitZoneObjects.
EmitZoneData takes over the role previously held by EmitZoneObject.