Removed interface EmitterOpRandomStartEndConfig
The interface EmitterOpRandomStartEndConfig, even though it is another path to configure a random emit value, it does not fully describe that all the options for the EmitterOpEaseConfig are available to use. Instead it is better to add to the description of the EmitterOpEaseConfig that there is a random start value option available.