Add basic RenderNode system. This currently only handles basic multi-textured images/sprites through a basic camera. It forms the basis of a major overhaul.
We assume everything is based on independent quads.
Our shaders are initialized with VAOs,
and drawn using instanced quads.
The buffer memory usage is about 1/6th previous.

Most render systems don't work, and WebGLRenderer is a mess.
That's OK; we just needed to get this committed before it got any
more complicated.