Fixing DomElement position when camera zoom != 1
When the camera zoom is set to any other value than one, if a dom element origin
if set to anything other than 0, the final position of the DOM was incorrect.

I traced this back to the DOMElementCSSRenderer.
After some trial and error, I realized the translation needed to be applied after camera multiplication
and not at the beginning.

With this change, DOMElements are correctly positioned.